Finally fixed that annoying texture issue

+++ CGP600-Assignment-02/include/BlockInstance.hpp
@@ -7,5 +7,5 @@
 struct BlockInstance
 {
     DirectX::XMFLOAT3 position;
-	float textureId;
+	UINT textureId;
 };
